requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_openmp_mutex=4.5=20_gnu
absl-py=2.4.0=pyhd8ed1ab_0
alsa-lib=1.2.15.3=hb03c661_0
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
asciitree=0.3.3=py_2
attr=2.5.2=hb03c661_1
brotli=1.2.0=hed03a55_1
brotli-bin=1.2.0=hb03c661_1
bzip2=1.0.8=hda65f42_9
c-ares=1.34.6=hb03c661_0
ca-certificates=2026.2.25=hbd8a1cb_0
cairo=1.18.4=he90730b_1
cloudpickle=3.1.2=pyhcf101f3_1
contourpy=1.3.2=py310h3788b33_0
cpython=3.10.20=py310hd8ed1ab_0
cuda-crt-tools=12.9.86=ha770c72_2
cuda-cudart=12.9.79=h5888daf_0
cuda-cudart_linux-64=12.9.79=h3f2d84a_0
cuda-cuobjdump=12.9.82=hffce074_1
cuda-cupti=12.9.79=h676940d_1
cuda-nvcc-tools=12.9.86=he02047a_2
cuda-nvdisasm=12.9.88=hffce074_1
cuda-nvrtc=12.9.86=hecca717_1
cuda-nvtx=12.9.79=hecca717_1
cuda-nvvm-tools=12.9.86=h4bc722e_2
cuda-version=12.9=h4f385c5_3
cudnn=9.10.2.21=hbcb9cd8_0
cusparselt=0.7.0.0=hcd2ec93_0
cycler=0.12.1=pyhcf101f3_2
cyrus-sasl=2.1.28=hac629b4_1
dbus=1.16.2=h24cb091_1
double-conversion=3.4.0=hecca717_0
drake=1.51.1=pypi_0
farama-notifications=0.0.4=pyhd8ed1ab_0
fasteners=0.19=pyhd8ed1ab_1
filelock=3.25.2=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.17.1=h27c8c51_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=hc364b38_1
fonttools=4.62.0=py310h3406613_0
freetype=2.14.2=ha770c72_0
fsspec=2026.2.0=pyhd8ed1ab_0
gmp=6.3.0=hac33072_2
gmpy2=2.3.0=py310h63ebcad_1
graphite2=1.3.14=hecca717_2
grpcio=1.68.2=py310h56e06c5_0
gymnasium=1.2.3=np2py310hf455f1b_0
harfbuzz=13.1.1=h6083320_0
icu=78.2=h33c6efd_0
importlib-metadata=8.7.0=pyhe01879c_1
jax-jumpy=1.0.0=pyhd8ed1ab_1
jinja2=3.1.6=pyhcf101f3_1
keyutils=1.6.3=hb9d3cd8_0
kiwisolver=1.5.0=py310haaf941d_0
krb5=1.22.2=ha1258a1_0
lcms2=2.18=h0c24ade_0
ld_impl_linux-64=2.45.1=default_hbd61a6d_101
lerc=4.1.0=hdb68285_0
libabseil=20240722.0=cxx17_hbbce691_4
libattr=2.5.2=hb03c661_1
libblas=3.11.0=5_h4a7cf45_openblas
libbrotlicommon=1.2.0=hb03c661_1
libbrotlidec=1.2.0=hb03c661_1
libbrotlienc=1.2.0=hb03c661_1
libcap=2.77=h3ff7636_0
libcblas=3.11.0=5_h0358290_openblas
libclang-cpp22.1=22.1.0=default_h99862b1_0
libclang13=22.1.0=default_h746c552_0
libcublas=12.9.1.4=h676940d_1
libcudnn=9.10.2.21=hf7e9902_0
libcudnn-dev=9.10.2.21=h58dd1b1_0
libcudss0=0.4.0.2=he55f5cd_2
libcufft=11.4.1.4=hecca717_1
libcufile=1.14.1.1=hbc026e6_1
libcups=2.3.3=h7a8fb5f_6
libcurand=10.3.10.19=h676940d_1
libcusolver=11.7.5.82=h676940d_2
libcusparse=12.5.10.65=hecca717_2
libdeflate=1.25=h17f619e_0
libdrm=2.4.125=hb03c661_1
libedit=3.1.20250104=pl5321h7949ede_0
libegl=1.7.0=ha4b6fd6_2
libexpat=2.7.4=hecca717_0
libffi=3.5.2=h3435931_0
libfreetype=2.14.2=ha770c72_0
libfreetype6=2.14.2=h73754d4_0
libgcc=15.2.0=he0feb66_18
libgcc-ng=15.2.0=h69a702a_18
libgfortran=15.2.0=h69a702a_18
libgfortran5=15.2.0=h68bc16d_18
libgl=1.7.0=ha4b6fd6_2
libglib=2.86.4=h6548e54_1
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=15.2.0=he0feb66_18
libgrpc=1.68.2=h25350d4_0
libiconv=1.18=h3b78370_2
libjpeg-turbo=3.1.2=hb03c661_0
liblapack=3.11.0=5_h47877c9_openblas
libllvm20=20.1.8=hf7376ad_1
libllvm22=22.1.1=hf7376ad_0
liblzma=5.8.2=hb03c661_0
libmagma=2.8.0=h566cb83_2
libnl=3.11.0=hb9d3cd8_0
libnsl=2.0.1=hb9d3cd8_1
libntlm=1.8=hb9d3cd8_0
libnvjitlink=12.9.86=hecca717_2
libopenblas=0.3.30=pthreads_h94d23a6_4
libopengl=1.7.0=ha4b6fd6_2
libpciaccess=0.18=hb9d3cd8_0
libpng=1.6.55=h421ea60_0
libpq=18.3=h9abb657_0
libprotobuf=5.28.3=h6128344_1
libre2-11=2024.07.02=hbbce691_2
libsqlite=3.52.0=hf4e2dac_0
libstdcxx=15.2.0=h934c35e_18
libstdcxx-ng=15.2.0=hdf11a46_18
libsystemd0=259.5=h6569c3e_0
libtiff=4.7.1=h9d88235_1
libtorch=2.6.0=cuda126_generic_h4a15719_200
libudev1=259.5=h6569c3e_0
libuuid=2.41.3=h5347b49_0
libuv=1.51.0=hb03c661_1
libvulkan-loader=1.4.341.0=h5279c79_0
libwebp-base=1.6.0=hd42ef1d_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.13.1=hca5e8e5_0
libxml2=2.15.2=he237659_0
libxml2-16=2.15.2=hca6bf5a_0
libxslt=1.1.43=h711ed8c_1
libzlib=1.3.1=hb9d3cd8_2
markdown=3.10.2=pyhcf101f3_0
markdown-it-py=4.0.0=pyhd8ed1ab_0
markupsafe=3.0.3=py310h3406613_1
matplotlib=3.10.8=py310hff52083_0
matplotlib-base=3.10.8=py310hfde16b3_0
mdurl=0.1.2=pyhd8ed1ab_1
mosek=11.1.2=pypi_0
mpc=1.3.1=h24ddda3_1
mpfr=4.2.2=he0a73b1_0
mpmath=1.4.1=pyhd8ed1ab_0
msgpack-python=1.1.2=py310h03d9f68_1
munkres=1.1.4=pyhd8ed1ab_1
nccl=2.29.3.1=h4d09622_0
ncurses=6.5=h2d0b736_3
networkx=3.4.2=pyh267e887_2
nomkl=1.0=h5ca1d4c_0
numcodecs=0.13.1=py310h5eaa309_0
numpy=2.2.6=py310hefbff90_0
omegaconf=2.3.0=pyhd8ed1ab_0
openjpeg=2.5.4=h55fea9a_0
openldap=2.6.10=hbde042b_1
openssl=3.6.1=h35e630c_1
optree=0.19.0=py310h03d9f68_0
packaging=26.0=pyhcf101f3_0
pandas=2.3.3=py310h0158d43_2
pcre2=10.47=haa7fec5_0
pillow=12.1.1=py310h5a73078_0
pip=26.0.1=pyh8b19718_0
pixman=0.46.4=h54a6638_1
protobuf=5.28.3=py310hf71b8c6_0
pthread-stubs=0.4=hb9d3cd8_1002
pybind11=3.0.2=pyh7a1b43c_0
pybind11-global=3.0.2=pyhc7ab6ef_0
pydot=4.0.1=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
pyparsing=3.3.2=pyhcf101f3_0
pyside6=6.10.2=py310hb9dbd67_1
python=3.10.20=h3c07f61_0_cpython
python-dateutil=2.9.0.post0=pyhe01879c_2
python-tzdata=2025.3=pyhd8ed1ab_0
python_abi=3.10=8_cp310
pytorch=2.6.0=cuda126_generic_py310_h9bb2754_200
pytz=2026.1.post1=pyhcf101f3_0
pyyaml=6.0.3=py310h3406613_1
qhull=2020.2=h434a139_5
qt6-main=6.10.2=h17e89b9_5
rdma-core=61.0=h192683f_0
re2=2024.07.02=h9925aae_2
readline=8.3=h853b02a_0
rich=14.3.3=pyhcf101f3_0
setuptools=82.0.1=pyh332efcf_0
six=1.17.0=pyhe01879c_1
sleef=3.9.0=ha0421bc_0
stable-baselines3=2.7.1=pyhd8ed1ab_1
sympy=1.14.0=pyh2585a3b_106
tensorboard=2.20.0=pyhe01879c_0
tensorboard-data-server=0.7.0=py310hf0f1799_4
tk=8.6.13=noxft_h366c992_103
tornado=6.5.4=py310ha78b2d2_0
tqdm=4.67.3=pyh8f84b5b_0
triton=3.2.0=cuda126py310h50ec074_1
typing-extensions=4.15.0=h396c80c_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025c=hc9c84f9_1
unicodedata2=17.0.1=py310h7c4b9e2_0
wayland=1.24.0=hd6090a7_1
werkzeug=3.1.6=pyhcf101f3_0
wheel=0.46.3=pyhd8ed1ab_0
xcb-util=0.4.1=h4f16b4b_2
xcb-util-cursor=0.1.6=hb03c661_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.47=hb03c661_0
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.6=he73a12e_0
xorg-libx11=1.8.13=he1eb515_0
xorg-libxau=1.0.12=hb03c661_1
xorg-libxcomposite=0.4.7=hb03c661_0
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb03c661_1
xorg-libxext=1.3.7=hb03c661_0
xorg-libxfixes=6.0.2=hb03c661_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.5=hb03c661_0
xorg-libxrender=0.9.12=hb9d3cd8_0
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.7=hb03c661_0
yaml=0.2.5=h280c20c_3
zarr=2.18.3=pyhd8ed1ab_1
zipp=3.23.0=pyhcf101f3_1
zlib-ng=2.3.3=hceb46e0_1
zstd=1.5.7=hb78ec9c_6
